[REVIEW] More DBSCAN fixes. Exposing both int and long variants through the API. Will need to find a good way to allow the user to select one through the Python API

+++ cpp/src/dbscan/adjgraph/naive.h
@@ -26,7 +26,7 @@ namespace Dbscan {
 namespace AdjGraph {
 namespace Naive {
 
-template <typename Type, typename Index_ = long>
+template <typename Type, typename Index_ = int>
 void launcher(const ML::cumlHandle_impl& handle, Pack<Type, Index_> data,
               Index_ batchSize, cudaStream_t stream) {
   Index_ k = 0;
